+++ example/project_template_example.dart
@@ -1,19 +1,26 @@
 import 'dart:io';
 
 import 'package:project_template/src/project_template_storage_io.dart';
+import 'package:yaml_writer/yaml_writer.dart';
 
 void main() async {
-  var currentDir = Directory.current;
-  var subDir = 'template-example';
+  var currentDir = Directory.current.absolute;
 
-  if (!currentDir.path.endsWith('example')) {
-    subDir = 'example/$subDir';
-  }
+  var exampleDir = currentDir.path.endsWith('example')
+      ? currentDir
+      : Directory('${currentDir.path}/example');
 
-  var storage = StorageIO.directory(currentDir, subDir)
+  assert(
+      exampleDir.existsSync(), "Can't resolve example directory: $exampleDir");
+
+  var templateDir = 'template-example';
+
+  var storage = StorageIO.directory(exampleDir, templateDir)
     ..ignoreFiles.add('.DS_Store');
 
   print(storage);
+
+  print('----------------------------------------------------');
 
   var files = storage.listFiles();
 
@@ -21,17 +28,43 @@ void main() async {
     print('- $file \t-> ${file.directoryAbsolute}');
   }
 
+  print('----------------------------------------------------');
+
   var template = await storage.loadTemplate();
 
   var templateVariables = template.parseTemplateVariables();
 
-  print('Variables: $templateVariables');
+  print('Variables:\n  $templateVariables\n');
+
+  var manifest = template.getManifest();
+
+  print('Manifest:\n');
+
+  print(
+    YAMLWriter()
+        .write(manifest)
+        .split(RegExp(r'[\r\n]'))
+        .map((l) => '  $l')
+        .join('\n'),
+  );
 
   print('----------------------------------------------------');
+
   print(template.toYAMLEncoded());
 
-  var templateResolved = template.resolve({'project_name': 'console_simple'});
-
   print('----------------------------------------------------');
+
+  var variables = {
+    'project_name': 'Console Simple',
+    'project_name_dir': 'console_simple',
+    'homepage': 'https://console-simple.domain',
+  };
+
+  var templateResolved = template.resolve(variables);
+
   print(templateResolved.toYAMLEncoded());
+
+  //// Save `templateResolved` to `example/template-generated`:
+  // var storageSave = StorageIO.directory(exampleDir);
+  // templateResolved.saveTo(storageSave, 'template-generated');
 }