+++ example/template-example/___project_name___/bin/___project_name___.dart
@@ -1,3 +1,0 @@
-void main(List<String> arguments) {
-  print('Project: ___project_name___');
-}