+++ example/template-example/___project_name_dir___/bin/___project_name_dir___.dart
@@ -1,0 +1,3 @@
+void main(List<String> arguments) {
+  print('Project[___project_name_dir___]: ___project_name___');
+}