+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0
 
+- CLI support.
+  - create, info, prepare. 
+- Template files support (YAML and JSON).
+- Template manifest support.
 - Initial version.